Handle backpressure in generated stream, add close hack
The VideoFrame stream generator should now correctly handle backpressure
signals. The timestamp it generates is no longer an ID but rather the
time at which the frame should be presented.

The timestamp encoded in the image is the timestamp in milliseconds.

This update implements an ugly postMessage hack to force Chrome to close
VideoFrame objects in workers, which in turn means it won't incorrectly
apply backpressure. The hack can be turned off by turning the videoHack
boolean to false in main.js