Enable overlay transform and re-factor code a bit
The overlay transform can now enabled or disabled through a radio button on the
test page. It produces a picture-in-picture with the color-encoded timestamp of
the frame. It can be enabled with generated VideoFrames as well (in which case
the picture-in-picture overlay is the same as the underlying frame).

The overlay is not yet used to extract the timestamp from the frame seen by
requestVideoFrameCallback().