Major re-write to add WebAssembly transformer
This adds a WebAssembly transformer that replaces any green present in a video
frame with W3C blue color.

The source code of the WebAssembly module is in `GreenBackgroundReplacer.wat`.
The code was written using the WebAssembly text format on purpose, as a way to
get hands-on with WebAssembly.

The binary code is generated with the `wat2wasm` package:
  https://www.npmjs.com/package/wat2wasm

... but the same result would be obtained using the online `wat2wasm` demo:
  https://webassembly.github.io/wabt/demo/wat2wasm/

The video background was switched to green to showcase the transformer.

To avoid having to handle the different video pixel formats in the WebAssembly
code, especially the ones that use a YUV arrangement, a new frame format
converter was added, leveraging WebGPU's sampler's ability to convert incoming
formats to RGBA/RGBX.

The chain of transformations was re-written to make it possible to select more
than one transform, as well as to add the timestamp overlay as the last
processing step (the green replacement transformation would have changed the
encoded timestamp otherwise).

The overlay was also downsized to be only 1/8 of the frame's dimensions.

User may now also select the video resolution that they want to generate.