Add ability to make explicit copies to CPU/GPU memory
New feature allows to study the impact of the frame's location on the following
processing step. User can choose to proceed as-is, to make explicit copies to
CPU memory, or to make explicit copies to GPU memory in between steps. Copy
statistics are reported as well when explicit copies are requested.